chore: dependencies updated, redis method updated

+++ src/user_io/ws_message.rs
@@ -9,7 +9,7 @@ pub mod wm {
     use ulid::Ulid;
 
     pub enum Error {
-        RateLimit(usize),
+        RateLimit(i64),
         InvalidStructure,
         MessageSize,
         MonthlyBandwidth,
